microservices/object-detection/pyproject.toml
[tool.poetry]
name = "object-detection"
version = "0.1.0"
description = "Object detection microservice"
authors = ["hfmartinez <heberthfabianmartinez@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.9.0.80"
fastapi = "^0.110.0"
sqlalchemy = "^2.0.28"
python-dotenv = "^1.0.1"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
uvicorn = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
